scheduler_agent/main.py
# main.py for scheduler_agent
from fastapi import FastAPI
from pydantic import BaseModel
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI()

class EmailData(BaseModel):
    # Define the structure of the email data
    pass

@app.post("/handle_schedule")
async def handle_schedule(email: EmailData):
    print("Handling scheduling request...")
    # TODO: Implement scheduling logic
    return {"status": "schedule handled"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8004)
